agent-openai-agents-sdk_2026_02_23-05_14/agent-openai-agents-sdk/pyproject.toml
[project]
name = "agent-openai-agents-sdk"
version = "0.1.0"
description = "Multi-agent SDK with PERFECT-AGENT (Nemotron fallback)"
readme = "README.md"
authors = [
    { name = "Agent Developer", email = "developer@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "fastapi",
    "uvicorn",
    "gradio",
    "httpx",
    "pytest",
    "python-dotenv",
    "mlflow>=2.0",
]

[project.optional-dependencies]
api = ["fastapi", "uvicorn"]
ui = ["gradio"]
databricks = ["databricks-openai>=0.9.0", "databricks-agents", "unitycatalog-openai[databricks]>=0.2.0"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "pytest>=7.0.0",
]


[project.scripts]
perfect-agent = "local_agents.perfect_agent.runner:main"
quickstart = "scripts.quickstart:main"
verify-setup = "verify_setup:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
discover-tools = "scripts.discover_tools:main"

[tool.setuptools.package-data]
"local_agents.perfect_agent" = ["system_prompt.txt"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["local_agents*", "api*", "ui*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
